Second Working Version (Good Night)

+++ Assignment2/our_sim/event.cc
@@ -1,6 +1,6 @@
 #include "common.hh"
 
-Event::Event(float timestamp, EventType type, Request* request, Thread* thread) {
+Event::Event(double timestamp, EventType type, Request* request, Thread* thread) {
     this->timestamp = timestamp;
     this->type = type;
     this->request = request;
